requirements.txt
aiohttp
backports.zoneinfo;python_version<"3.9"
rsa
